Create network graph component
